Change updateScopeRuns.m to ingest csv
Previous version took in the plain webpage and parsed it; this edit requests a csv from the Google API, for a (hopefully) more robust result.

It still takes in the URL as before (so as to not change the user experience). However, the URL should be for a regular publish, rather than a 'publish as webpage'. Documentation should be updated accordingly.

Because we are now requesting a csv, we no longer need to decode HTML &#39; occurrences.